+++ src/Events/Events.js
@@ -183,9 +183,10 @@ class Events extends PureComponent {
     },
   );
 
-  onGridClick = (event, dayIndex) => {
-    const { initialDate, onGridClick } = this.props;
-    if (!onGridClick) {
+  onGridTouch = (event, dayIndex, longPress) => {
+    const { initialDate, onGridClick, onGridLongPress } = this.props;
+    const callback = longPress ? onGridLongPress : onGridClick;
+    if (!callback) {
       return;
     }
     const { locationY } = event.nativeEvent;
@@ -193,7 +194,7 @@ class Events extends PureComponent {
 
     const date = moment(initialDate).add(dayIndex, 'day').toDate();
 
-    onGridClick(event, hour, date);
+    callback(event, hour, date);
   };
 
   isToday = (dayIndex) => {
@@ -209,6 +210,7 @@ class Events extends PureComponent {
       numberOfDays,
       times,
       onEventPress,
+      onEventLongPress,
       eventContainerStyle,
       EventComponent,
       rightToLeft,
@@ -240,7 +242,8 @@ class Events extends PureComponent {
         <View style={styles.events}>
           {totalEvents.map((eventsInSection, dayIndex) => (
             <TouchableWithoutFeedback
-              onPress={(e) => this.onGridClick(e, dayIndex)}
+              onPress={(e) => this.onGridTouch(e, dayIndex, false)}
+              onLongPress={(e) => this.onGridTouch(e, dayIndex, true)}
               key={dayIndex}
             >
               <View style={styles.event}>
@@ -257,6 +260,7 @@ class Events extends PureComponent {
                     event={item.data}
                     position={item.style}
                     onPress={onEventPress}
+                    onLongPress={onEventLongPress}
                     EventComponent={EventComponent}
                     containerStyle={eventContainerStyle}
                   />
@@ -279,7 +283,9 @@ Events.propTypes = {
   timeStep: PropTypes.number.isRequired,
   times: PropTypes.arrayOf(PropTypes.string).isRequired,
   onEventPress: PropTypes.func,
+  onEventLongPress: PropTypes.func,
   onGridClick: PropTypes.func,
+  onGridLongPress: PropTypes.func,
   eventContainerStyle: PropTypes.object,
   EventComponent: PropTypes.elementType,
   rightToLeft: PropTypes.bool,